+++ xudt/Cargo.toml
@@ -15,6 +15,5 @@ ckb-script = "0.40.0"
 ckb-traits = "0.40.0"
 ckb-types = "0.40.0"
 lazy_static = "1.3.0"
-molecule = { version = "0.7", default-features = false }
 rand = "0.6.5"
 sparse-merkle-tree = "0.4.0-rc1"
