+++ xudt/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod blockchain;
 pub mod xudt_rce_mol;
