+++ xudt/src/xudt_rce_mol.rs
@@ -1,7 +1,13 @@
 // Generated by Molecule 0.7.0
 
-use super::blockchain::*;
-use molecule::prelude::*;
+#![allow(unused_imports)]
+
+use ckb_types::prelude::*;
+use ckb_types::molecule;
+use ckb_types::packed::*;
+// these lines above are manually added
+// replace "::molecule" to "molecule" in below code
+
 #[derive(Clone)]
 pub struct ScriptVec(molecule::bytes::Bytes);
 impl ::core::fmt::LowerHex for ScriptVec {
@@ -253,7 +259,7 @@ impl molecule::prelude::Builder for ScriptVecBuilder {
                 .map(|inner| inner.as_slice().len())
                 .sum::<usize>()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let item_count = self.0.len();
         if item_count == 0 {
             writer.write_all(&molecule::pack_number(
@@ -486,7 +492,7 @@ impl molecule::prelude::Builder for ScriptVecOptBuilder {
             .map(|ref inner| inner.as_slice().len())
             .unwrap_or(0)
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         self.0
             .as_ref()
             .map(|ref inner| writer.write_all(inner.as_slice()))
@@ -739,7 +745,7 @@ impl molecule::prelude::Builder for XudtWitnessInputBuilder {
             + self.raw_extension_data.as_slice().len()
             + self.extension_data.as_slice().len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
         let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
         offsets.push(total_size);
@@ -915,7 +921,7 @@ impl molecule::prelude::Builder for RCRuleBuilder {
     fn expected_length(&self) -> usize {
         Self::TOTAL_SIZE
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(self.smt_root.as_slice())?;
         writer.write_all(self.flags.as_slice())?;
         Ok(())
@@ -1130,7 +1136,7 @@ impl molecule::prelude::Builder for RCCellVecBuilder {
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.0.len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(&molecule::pack_number(self.0.len() as molecule::Number))?;
         for inner in &self.0[..] {
             writer.write_all(inner.as_slice())?;
@@ -1350,7 +1356,7 @@ impl molecule::prelude::Builder for RCDataBuilder {
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE + self.0.as_slice().len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(&molecule::pack_number(self.0.item_id()))?;
         writer.write_all(self.0.as_slice())
     }
@@ -1687,7 +1693,7 @@ impl molecule::prelude::Builder for SmtProofBuilder {
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.0.len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(&molecule::pack_number(self.0.len() as molecule::Number))?;
         for inner in &self.0[..] {
             writer.write_all(inner.as_slice())?;
@@ -1965,7 +1971,7 @@ impl molecule::prelude::Builder for SmtProofEntryBuilder {
             + self.mask.as_slice().len()
             + self.proof.as_slice().len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
         let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
         offsets.push(total_size);
@@ -2238,7 +2244,7 @@ impl molecule::prelude::Builder for SmtProofEntryVecBuilder {
                 .map(|inner| inner.as_slice().len())
                 .sum::<usize>()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let item_count = self.0.len();
         if item_count == 0 {
             writer.write_all(&molecule::pack_number(
@@ -2475,7 +2481,7 @@ impl molecule::prelude::Builder for SmtUpdateItemBuilder {
     fn expected_length(&self) -> usize {
         Self::TOTAL_SIZE
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(self.key.as_slice())?;
         writer.write_all(self.packed_values.as_slice())?;
         Ok(())
@@ -2690,7 +2696,7 @@ impl molecule::prelude::Builder for SmtUpdateItemVecBuilder {
     fn expected_length(&self) -> usize {
         molecule::NUMBER_SIZE + Self::ITEM_SIZE * self.0.len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         writer.write_all(&molecule::pack_number(self.0.len() as molecule::Number))?;
         for inner in &self.0[..] {
             writer.write_all(inner.as_slice())?;
@@ -2995,7 +3001,7 @@ impl molecule::prelude::Builder for SmtUpdateActionBuilder {
             + self.updates.as_slice().len()
             + self.proof.as_slice().len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
         let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
         offsets.push(total_size);
@@ -3257,7 +3263,7 @@ impl molecule::prelude::Builder for XudtDataBuilder {
             + self.lock.as_slice().len()
             + self.data.as_slice().len()
     }
-    fn write<W: ::molecule::io::Write>(&self, writer: &mut W) -> ::molecule::io::Result<()> {
+    fn write<W: molecule::io::Write>(&self, writer: &mut W) -> molecule::io::Result<()> {
         let mut total_size = molecule::NUMBER_SIZE * (Self::FIELD_COUNT + 1);
         let mut offsets = Vec::with_capacity(Self::FIELD_COUNT);
         offsets.push(total_size);
